+++ src/qt/sendcoinsentry.cpp
@@ -36,7 +36,8 @@ SendCoinsEntry::SendCoinsEntry(const PlatformStyle *_platformStyle,
         tr("A message that was attached to the %1 URI which will be"
            " stored with the transaction for your reference. Note: "
            "This message will not be sent over the Bitcoin network.")
-            .arg(GUIUtil::bitcoinURIScheme(GetConfig())));
+            .arg(QString::fromStdString(
+                model->getChainParams().CashAddrPrefix())));
 
     setCurrentWidget(ui->SendCoins);