+++ src/qt/test/uritests.h
@@ -15,7 +15,6 @@ private Q_SLOTS:
     void uriTestsBase58();
     void uriTestsCashAddr();
     void uriTestFormatURI();
-    void uriTestScheme();
 };
 
 #endif // BITCOIN_QT_TEST_URITESTS_H